+++ src/omnibar/omnibar-config.ts
@@ -60,5 +60,4 @@ export interface BBOmnibarConfig {
   navVersion?: string;
 
   hideResourceLinks?: boolean;
-
 }