+++ src/omnibar/omnibar-toast-container-init-args.ts
@@ -1,0 +1,25 @@
+//#region imports
+
+import {
+  BBOmnibarNavigationItem
+} from './omnibar-navigation-item';
+
+//#endregion
+
+export interface BBOmnibarToastContainerInitArgs {
+
+  envId: string;
+
+  leId: string;
+
+  openMenuCallback: () => void;
+
+  navigateCallback: (navItem: BBOmnibarNavigationItem) => void;
+
+  navigateUrlCallback: (url: string) => void;
+
+  svcId: string;
+
+  url: string;
+
+}