+++ src/omnibar/omnibar.ts
@@ -1,5 +1,7 @@
 //#region imports
 
+import * as jwtDecode from 'jwt-decode';
+
 import {
   BBAuth
 } from '../auth';
@@ -7,6 +9,14 @@ import {
 import {
   BBAuthInterop
 } from '../shared/interop';
+
+import {
+  BBAuthNavigator
+} from '../shared/navigator';
+
+import {
+  BBAuthDomUtility
+} from '../shared/dom-utility';
 
 import {
   BBOmnibarConfig
@@ -25,6 +35,10 @@ import {
 } from './omnibar-search-args';
 
 import {
+  BBOmnibarSetTitleArgs
+} from './omnibar-set-title-args';
+
+import {
   BBOmnibarUserActivity
 } from './omnibar-user-activity';
 
@@ -37,12 +51,12 @@ import {
 } from './omnibar-update-args';
 
 import {
-  BBAuthNavigator
-} from '../shared/navigator';
+  BBOmnibarPushNotifications
+} from './omnibar-push-notifications';
 
 import {
-  BBAuthDomUtility
-} from '../shared/dom-utility';
+  BBOmnibarToastContainer
+} from './omnibar-toast-container';
 
 import {
   BBOmnibarThemeAccent
@@ -60,8 +74,12 @@ let iframeEl: HTMLIFrameElement;
 let omnibarConfig: BBOmnibarConfig;
 let currentLegacyKeepAliveUrl: string;
 let promiseResolve: () => void;
+let pushNotificationsConnected: boolean;
+let unreadNotificationCount: number;
+let serviceName: string;
+let currentTitleParts: string[];
 
-function addIframeEl() {
+function addIframeEl(): void {
   iframeEl = BBAuthDomUtility.addIframe(
     buildOmnibarUrl(),
     `sky-omnibar-iframe ${CLS_LOADING}`,
@@ -69,18 +87,18 @@ function addIframeEl() {
   );
 }
 
-function addEnvironmentEl() {
+function addEnvironmentEl(): void {
   envEl = document.createElement('div');
   envEl.className = 'sky-omnibar-environment';
 
   BBAuthDomUtility.addElToBodyTop(envEl);
 }
 
-function collapseIframe() {
+function collapseIframe(): void {
   iframeEl.classList.remove(CLS_EXPANDED);
 }
 
-function addStyleEl() {
+function addStyleEl(): void {
   let accentCss = 'background: linear-gradient(to right, #71bf44 0, #31b986 50%, #00b2ec 100%);';
   let backgroundColor = '#4d5259';
 
@@ -131,15 +149,15 @@ body {
   ${accentCss}
 }
 
-.sky-omnibar-placeholder.sky-omnibar-loading {
+.sky-omnibar-placeholder.${CLS_LOADING} {
   display: block;
 }
 
-.sky-omnibar-iframe.sky-omnibar-loading {
+.sky-omnibar-iframe.${CLS_LOADING} {
   visibility: hidden;
 }
 
-.sky-omnibar-iframe-expanded {
+.${CLS_EXPANDED} {
   height: 100%;
 }
 
@@ -165,7 +183,7 @@ body {
   );
 }
 
-function addPlaceholderEl() {
+function addPlaceholderEl(): void {
   placeholderEl = document.createElement('div');
   placeholderEl.className = `sky-omnibar-placeholder ${CLS_LOADING}`;
   placeholderEl.innerHTML = `<div class="sky-omnibar-placeholder-accent"></div>`;
@@ -173,21 +191,25 @@ function addPlaceholderEl() {
   document.body.appendChild(placeholderEl);
 }
 
-function expandIframe() {
+function expandIframe(): void {
   iframeEl.classList.add(CLS_EXPANDED);
 }
 
-function handleStateChange() {
+function handleStateChange(): void {
+  const url = document.location.href;
+
   BBAuthInterop.postOmnibarMessage(
     iframeEl,
     {
-      href: document.location.href,
+      href: url,
       messageType: 'location-change'
     }
   );
+
+  BBOmnibarToastContainer.updateUrl(url);
 }
 
-function handleSearch(searchArgs: BBOmnibarSearchArgs) {
+function handleSearch(searchArgs: BBOmnibarSearchArgs): void {
   if (omnibarConfig.onSearch) {
     omnibarConfig
       .onSearch(searchArgs)
@@ -203,7 +225,88 @@ function handleSearch(searchArgs: BBOmnibarSearchArgs) {
   }
 }
 
-function refreshUserCallback() {
+function openPushNotificationsMenu(): void {
+  BBAuthInterop.postOmnibarMessage(
+    iframeEl,
+    {
+      messageType: 'push-notifications-open'
+    }
+  );
+}
+
+function hasNotificationsEntitlement(token: string): boolean {
+  const decodedToken: any = jwtDecode(token);
+  let entitlements: string | string[] = decodedToken['1bb.entitlements'];
+
+  if (entitlements) {
+    entitlements = Array.isArray(entitlements) ? entitlements : [entitlements];
+    return (entitlements as string[]).indexOf('notif') > -1;
+  }
+
+  return false;
+}
+
+function connectPushNotifications(): void {
+  if (!pushNotificationsConnected) {
+    pushNotificationsConnected = true;
+
+    BBAuth.getToken({
+      disableRedirect: true,
+      envId: omnibarConfig.envId,
+      leId: omnibarConfig.leId,
+      permissionScope: 'Notifications'
+    }).then((token: string) => {
+      if (hasNotificationsEntitlement(token)) {
+        BBOmnibarToastContainer.init({
+          envId: omnibarConfig.envId,
+          leId: omnibarConfig.leId,
+          navigateCallback: handleNavigate,
+          navigateUrlCallback: handleNavigateUrl,
+          openMenuCallback: openPushNotificationsMenu,
+          svcId: omnibarConfig.svcId,
+          url: document.location.href
+        })
+          .then(() => {
+            BBOmnibarPushNotifications.connect(
+              omnibarConfig.leId,
+              omnibarConfig.envId,
+              (notifications) => {
+                BBAuthInterop.postOmnibarMessage(
+                  iframeEl,
+                  {
+                    messageType: 'push-notifications-update',
+                    pushNotifications: notifications
+                  }
+                );
+
+                BBOmnibarToastContainer.showNewNotifications(notifications);
+
+                unreadNotificationCount = notifications &&
+                  notifications.notifications &&
+                  notifications.notifications.filter((notification: any) => !notification.isRead).length;
+
+                updateTitle();
+              });
+            });
+      } else {
+        pushNotificationsConnected = false;
+      }
+    }).catch(() => {
+      pushNotificationsConnected = false;
+    });
+  }
+}
+
+function disconnectPushNotifications(): void {
+  if (pushNotificationsConnected) {
+    BBOmnibarToastContainer.destroy();
+    BBOmnibarPushNotifications.disconnect();
+
+    pushNotificationsConnected = false;
+  }
+}
+
+function refreshUserCallback(): void {
   function refreshUser(token: string) {
     BBAuthInterop.postOmnibarMessage(
       iframeEl,
@@ -212,6 +315,12 @@ function refreshUserCallback() {
         token
       }
     );
+
+    if (token) {
+      connectPushNotifications();
+    } else {
+      disconnectPushNotifications();
+    }
   }
 
   BBAuth.clearTokenCache();
@@ -224,7 +333,7 @@ function refreshUserCallback() {
     .catch(() => refreshUser(undefined));
 }
 
-function showInactivityCallback() {
+function showInactivityCallback(): void {
   BBOmnibarUserActivityPrompt.show({
     sessionRenewCallback: () => {
       BBOmnibarUserActivity.userRenewedSession();
@@ -232,11 +341,11 @@ function showInactivityCallback() {
   });
 }
 
-function hideInactivityCallback() {
+function hideInactivityCallback(): void {
   BBOmnibarUserActivityPrompt.hide();
 }
 
-function startActivityTracking() {
+function startActivityTracking(): void {
   BBOmnibarUserActivity.startTracking(
     refreshUserCallback,
     showInactivityCallback,
@@ -249,7 +358,7 @@ function startActivityTracking() {
 function handleGetToken(
   tokenRequestId: any,
   disableRedirect: boolean
-) {
+): void {
   BBAuth.getToken({
     disableRedirect
   })
@@ -281,7 +390,7 @@ function handleGetToken(
     );
 }
 
-function handleHelp() {
+function handleHelp(): void {
   const BBHELP = (window as any).BBHELP;
 
   if (BBHELP) {
@@ -289,12 +398,16 @@ function handleHelp() {
   }
 }
 
-function handleNotificationRead(notification: BBOmnibarNotificationItem) {
+function handleNotificationRead(notification: BBOmnibarNotificationItem): void {
   const notificationsConfig = omnibarConfig.notifications;
 
   if (notificationsConfig && notificationsConfig.onNotificationRead) {
     notificationsConfig.onNotificationRead(notification);
   }
+}
+
+function handlePushNotificationsChange(notifications: any[]): void {
+  BBOmnibarPushNotifications.updateNotifications(notifications);
 }
 
 function handleEnvironmentUpdate(name: string) {
@@ -312,7 +425,19 @@ function handleEnvironmentUpdate(name: string) {
   }
 }
 
-function monkeyPatchState() {
+function handleNavigate(navItem: BBOmnibarNavigationItem): void {
+  const nav = omnibarConfig.nav;
+
+  if (!nav || !nav.beforeNavCallback || nav.beforeNavCallback(navItem) !== false) {
+    BBAuthNavigator.navigate(navItem.url);
+  }
+}
+
+function handleNavigateUrl(url: string): void {
+  BBAuthNavigator.navigate(url);
+}
+
+function monkeyPatchState(): void {
   const oldPushState = history.pushState;
   const oldReplaceState = history.replaceState;
 
@@ -336,7 +461,7 @@ function monkeyPatchState() {
   history.replaceState = newReplaceState;
 }
 
-function setupNotifications() {
+function initLocalNotifications(): void {
   const notificationsConfig = omnibarConfig.notifications;
 
   if (notificationsConfig) {
@@ -354,7 +479,7 @@ function setupNotifications() {
   }
 }
 
-function messageHandler(event: MessageEvent) {
+function messageHandler(event: MessageEvent): void {
   if (!BBAuthInterop.messageIsFromOmnibar(event)) {
     return;
   }
@@ -397,7 +522,8 @@ function messageHandler(event: MessageEvent) {
         }
       );
 
-      setupNotifications();
+      initLocalNotifications();
+      connectPushNotifications();
 
       handleStateChange();
 
@@ -414,15 +540,10 @@ function messageHandler(event: MessageEvent) {
       collapseIframe();
       break;
     case 'navigate-url':
-      BBAuthNavigator.navigate(message.url);
+      handleNavigateUrl(message.url);
       break;
     case 'navigate':
-      const navItem: BBOmnibarNavigationItem = message.navItem;
-
-      if (!nav || !nav.beforeNavCallback || nav.beforeNavCallback(navItem) !== false) {
-        BBAuthNavigator.navigate(navItem.url);
-      }
-
+      handleNavigate(message.navItem);
       break;
     case 'search':
       handleSearch(message.searchArgs);
@@ -441,6 +562,9 @@ function messageHandler(event: MessageEvent) {
         message.notification
       );
       break;
+    case 'push-notifications-change':
+      handlePushNotificationsChange(message.notifications);
+      break;
     case 'session-renew':
       BBOmnibarUserActivity.userRenewedSession();
       break;
@@ -451,15 +575,37 @@ function messageHandler(event: MessageEvent) {
       currentLegacyKeepAliveUrl = message.url;
       startActivityTracking();
       break;
+
+    case 'selected-service-update':
+      serviceName = message.serviceName;
+      updateTitle();
   }
 }
 
-function buildOmnibarUrl() {
+function buildOmnibarUrl(): string {
   const omnibarUrl = omnibarConfig.url ||
     /* istanbul ignore next */
     'https://app.blackbaud.com/omnibar/';
 
   return omnibarUrl;
+}
+
+function updateTitle(): void {
+  if (currentTitleParts) {
+    const titleParts = currentTitleParts.slice();
+
+    if (serviceName) {
+      titleParts.push(serviceName);
+    }
+
+    let title = titleParts.join(' - ');
+
+    if (unreadNotificationCount) {
+      title = `(${unreadNotificationCount}) ${title}`;
+    }
+
+    document.title = title;
+  }
 }
 
 export class BBOmnibar {
@@ -485,7 +631,7 @@ export class BBOmnibar {
     });
   }
 
-  public static update(args: BBOmnibarUpdateArgs) {
+  public static update(args: BBOmnibarUpdateArgs): void {
     BBAuthInterop.postOmnibarMessage(
       iframeEl,
       {
@@ -495,12 +641,20 @@ export class BBOmnibar {
     );
   }
 
-  public static destroy() {
-    BBAuthDomUtility.removeCss(styleEl);
+  public static setTitle(args: BBOmnibarSetTitleArgs): void {
+    currentTitleParts = args && args.titleParts;
+    updateTitle();
+  }
+
+  public static destroy(): void {
+    BBOmnibarToastContainer.destroy();
+    BBOmnibarPushNotifications.disconnect();
 
     BBAuthDomUtility.removeEl(placeholderEl);
     BBAuthDomUtility.removeEl(iframeEl);
     BBAuthDomUtility.removeEl(envEl);
+
+    BBAuthDomUtility.removeCss(styleEl);
 
     window.removeEventListener('message', messageHandler);
 
@@ -510,6 +664,10 @@ export class BBOmnibar {
       iframeEl =
       envEl =
       promiseResolve =
+      pushNotificationsConnected =
+      unreadNotificationCount =
+      currentTitleParts =
+      serviceName =
       undefined;
   }
 }